Refactor: split login into its own file, split common response in its own file.
