+++ src/web/users/responses.rs
@@ -1,0 +1,7 @@
+use crate::models::User;
+use serde::Serialize;
+
+#[derive(Serialize, Debug)]
+pub struct UserResponse {
+    pub user: User,
+}
