+++ src/web/users/users.rs
@@ -1,11 +1,10 @@
-use crate::auth::encode_token;
 use crate::conduit::users;
 use crate::middleware::ContextExt;
 use crate::models::*;
 use crate::web::diesel_error;
 use crate::Repo;
 use log::info;
-use serde::{Deserialize, Serialize};
+use serde::Deserialize;
 
 use http::status::StatusCode;
 use tide::{Request, Response};
@@ -13,42 +12,6 @@ use tide::{Request, Response};
 #[derive(Deserialize, Debug)]
 pub struct UpdateUserRequest {
     user: UpdateUser,
-}
-
-#[allow(dead_code)]
-#[derive(Serialize)]
-pub struct UserResponse {
-    user: User,
-}
-
-#[derive(Deserialize)]
-pub struct AuthRequest {
-    user: AuthUser,
-}
-
-#[derive(Deserialize)]
-pub struct AuthUser {
-    email: String,
-    password: String,
-}
-
-pub async fn login(mut cx: Request<Repo>) -> tide::Result<Response> {
-    let auth: AuthRequest = cx.body_json().await.map_err(|_| StatusCode::BAD_REQUEST)?;
-    let repo = cx.state();
-    let user = auth.user;
-    let result = users::find_by_email_password(repo, user.email, user.password);
-
-    match result {
-        Ok(user) => {
-            let user = User {
-                token: Some(encode_token(user.id)),
-                ..user
-            };
-            Ok(Response::new(200).body_json(&user).unwrap())
-        }
-        Err(diesel::result::Error::NotFound) => Err(StatusCode::UNAUTHORIZED.into()),
-        Err(e) => Err(diesel_error(&e)),
-    }
 }
 
 pub async fn get_user(cx: Request<Repo>) -> tide::Result<Response> {
